Move withCluster and other test functions into their own module
