+++ lib/shelley/cardano-wallet.cabal
@@ -98,6 +98,7 @@ library
       Cardano.Wallet.Shelley.Network
       Cardano.Wallet.Shelley.Transaction
       Cardano.Wallet.Shelley.Launch
+      Cardano.Wallet.Shelley.Launch.Cluster
       Cardano.Wallet.Shelley.Pools
 
 executable cardano-wallet
